@@ -2,8 +2,8 @@
   "repository": "sharetribe/sharetribe",
   "license": "MIT",
   "engines": {
-    "node": "10.15.3",
-    "npm": "6.4.1"
+    "node": "18.16.0",
+    "npm": "9.5.1"
   },
   "scripts": {
     "ensure-node-version": "check-node-version --package",
@@ -40,78 +40,75 @@
     "@babel/polyfill": "7.4.4",
     "@babel/preset-env": "7.4.4",
     "@babel/preset-react": "7.0.0",
+    "@babel/register": "7.4.4",
     "@babel/runtime": "7.4.4",
-    "axios": "^0.21.2",
+    "axios": "^0.21.4",
     "babel-loader": "8.2.5",
     "check-node-version": "3.3.0",
     "classnames": "2.2.5",
-    "css-loader": "3.0.0",
+    "css-loader": "4.3.0",
     "es6-shim": "0.35.6",
-    "file-loader": "3.0.1",
+    "file-loader": "5.1.0",
     "i18n-js": "http://github.com/sharetribe/i18n-js/archive/interpolation-mode.tar.gz",
     "immutable": "3.8.1",
     "imports-loader": "0.6.5",
-    "loader-utils": "0.2.16",
+    "loader-utils": "^3.2.1",
     "lodash": "^4.17.21",
     "lodash.merge": "^4.6.2",
-    "mini-css-extract-plugin": "0.6.0",
-    "moment": "2.29.2",
+    "mini-css-extract-plugin": "1.6.2",
+    "moment": "^2.29.4",
     "node-libs-browser": "2.0.0",
-    "node-sass": "4.13.1",
     "numbro": "1.9.3",
     "postcss-cssnext": "3.1.0",
-    "postcss-loader": "3.0.0",
-    "postcss-mixins": "6.2.1",
-    "r-dom": "2.3.1",
-    "raw-loader": "2.0.0",
-    "react": "15.4.1",
-    "react-addons-css-transition-group": "15.4.1",
-    "react-addons-shallow-compare": "15.4.1",
-    "react-addons-transition-group": "15.4.1",
-    "react-dates": "4.1.0",
-    "react-dom": "15.4.1",
+    "postcss-loader": "5.3.0",
+    "postcss-mixins": "^9.0.4",
+    "prop-types": "15.8.1",
+    "r-dom": "2.4.0",
+    "raw-loader": "4.0.2",
+    "react": "16.1.1",
+    "react-addons-shallow-compare": "16.0.0-alpha.3",
+    "react-addons-transition-group": "16.0.0-alpha.3",
+    "react-dates": "16.7.1",
+    "react-dom": "16.1.1",
     "react-form": "git+https://github.com/ithouse/react-form.git#8053d0c9093c35843522b734fb5a1777b35f1c48",
     "react-on-rails": "13.0.2",
-    "react-redux": "4.4.6",
+    "react-redux": "5.1.2",
+    "react-transition-group": "2.9.0",
     "redux": "3.6.0",
     "redux-thunk": "2.1.0",
-    "resolve-url-loader": "1.6.0",
-    "sass-loader": "7.1.0",
+    "resolve-url-loader": "^5.0.0",
+    "sass": "1.32.13",
+    "sass-loader": "8.0.2",
     "style-loader": "0.23.1",
     "transit-js": "0.8.846",
-    "uglifyjs-webpack-plugin": "2.2.0",
-    "url-loader": "1.1.2",
-    "webpack": "4.30.0",
-    "webpack-cli": "3.3.0",
+    "url-loader": "3.0.0",
+    "webpack": "5.84.1",
+    "webpack-cli": "^5.1.1",
     "whatwg-fetch": "3.6.2"
   },
   "devDependencies": {
     "babel-eslint": "10.0.1",
     "babel-plugin-react-transform": "3.0.0",
-    "@babel/register": "7.4.4",
     "chai": "3.5.0",
-    "chai-enzyme": "0.6.1",
-    "enzyme": "2.6.0",
-    "eslint": "4.18.2",
+    "chai-enzyme": "1.0.0-beta.1",
+    "enzyme": "^3.5.0",
+    "enzyme-adapter-react-16": "^1.15.7",
+    "eslint": "5.16.0",
     "eslint-plugin-babel": "4.1.2",
-    "eslint-plugin-react": "6.4.1",
+    "eslint-plugin-react": "7.12.4",
     "ignore-styles": "5.0.1",
-    "jsdom": "9.8.3",
+    "jsdom": "^22.0.0",
     "json": "10.0.0",
     "mocha": "6.1.4",
-    "mocha-circleci-reporter": "0.0.3",
+    "mocha-circleci-reporter": "^0.0.2",
     "phantomjs-prebuilt": "^2.1.16",
-    "react-addons-test-utils": "15.4.1",
+    "react-addons-test-utils": "16.0.0-alpha.3",
     "react-transform-hmr": "1.0.4",
-    "@storybook/react": "^4.1.18",
-    "@storybook/addon-actions": "^4.1.18",
-    "@storybook/addon-links": "^4.1.18",
-    "@storybook/addons": "^4.1.18",
     "storybook-addon-specifications": "2.1.2",
     "stylelint": "10.0.1",
     "stylelint-config-standard": "18.3.0",
-    "webpack-dev-middleware": "3.6.1",
-    "webpack-dev-server": "3.3.1"
+    "webpack-dev-middleware": "4.3.0",
+    "webpack-dev-server": "4.15.0"
   },
   "browser": {
     "fs": false
