@@ -1,4 +1,5 @@
-import React, { Component, PropTypes } from 'react';
+import React, { Component } from 'react';
+import PropTypes from 'prop-types';
 import { Form, StyledSelect, FormField, Checkbox } from 'react-form';
 import merge from 'lodash.merge';
 import { t } from '../../../utils/i18n';
@@ -138,7 +139,7 @@ class DayWrapper extends Component {
     return { remove: !this.props.enabled };
   }
 
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     this.setState({ // eslint-disable-line react/no-set-state
       timeSlots: nextProps.timeSlots,
       enabled: nextProps.enabled, // eslint-disable-line no-underscore-dangle
@@ -250,7 +251,7 @@ class ListingWorkingHoursForm extends Component {
     this.formApi = null;
   }
 
-  componentWillReceiveProps(nextProps) {
+  UNSAFE_componentWillReceiveProps(nextProps) {
     if (this.formApi && nextProps.saveFinished) {
       this.formApi.setAllValues(nextProps.workingTimeSlots);
     }
