@@ -1,4 +1,5 @@
-import { Component, PropTypes } from 'react';
+import { Component } from 'react';
+import PropTypes from 'prop-types';
 import r, { a, div, img, span } from 'r-dom';
 import classNames from 'classnames';
 import { t, fullLocaleCode, localizedString, localizedPricingUnit } from '../../../utils/i18n';
