@@ -1,4 +1,5 @@
-import { Component, PropTypes } from 'react';
+import { Component } from 'react';
+import PropTypes from 'prop-types';
 import r, { div, span } from 'r-dom';
 import Link from '../../elements/Link/Link';
 import css from './MenuMobile.css';
