@@ -1,4 +1,5 @@
-import { Component, PropTypes } from 'react';
+import { Component } from 'react';
+import PropTypes from 'prop-types';
 import { div, span, a } from 'r-dom';
 import { t } from '../../../utils/i18n';
 
