@@ -1,4 +1,5 @@
-import { Component, PropTypes } from 'react';
+import { Component } from 'react';
+import PropTypes from 'prop-types';
 import r, { div } from 'r-dom';
 import Immutable from 'immutable';
 import classNames from 'classnames';
